requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
dj-database-url==3.1.2
Django==5.2.7
djangorestframework==3.16.1
gunicorn==25.3.0
idna==3.11
packaging==26.1
psycopg2-binary==2.9.11
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.12.0
